#495 (partial): extract submission output-formatting helpers
WebRoutes+Submission.swift held ~290 lines of pure stdout/stderr
parsing and chickadee.py JSON-envelope formatting helpers, plus the
SubmitFormBody multipart shape. None of these depend on Vapor routing
machinery — they're pure functions over strings + the TestStatus enum.

Moved to Sources/APIServer/Helpers/SubmissionOutputFormatting.swift.
WebRoutes+Submission.swift goes 854 → 568 LOC. No behaviour change,
all visibility preserved (file-private helpers stay file-private in
the new file).

Scope note: the issue also called for splitting AssignmentRoutes+NewAssignment.swift
(842 LOC) and AdminRoutes.swift (776 LOC). Both have tight clusters of
`private` extension helpers shared across adjacent route handlers in
the same file. Splitting either would force a visibility regression
to `internal` purely for LOC reduction — not worth it. The original
AssignmentRoutes.swift (1,401 LOC at the time #495 was filed) has
already been split into 17 sibling files; #495 was filed against a
stale snapshot.
